move function from src to test function where it belongs
